chore: rename integration status component
