+++ packages/app/src/client/viewsBasic/tenantIntegrations/tabs/Installed.tsx
@@ -37,7 +37,7 @@ import TableCell from "@material-ui/core/TableCell";
 import TableContainer from "@material-ui/core/TableContainer";
 import TableHead from "@material-ui/core/TableHead";
 import TableRow from "@material-ui/core/TableRow";
-import { IntegrationStatus } from "client/viewsBasic/integrationDefs/k8sMetrics/IntegrationStatus";
+import { IntegrationStatus } from "client/viewsBasic/integrationDefs/k8sMetrics/Status";
 
 const useStyles = makeStyles(theme => ({
   integrationRow: {
